Fix Timetabling.Tests, build setup, AppVeyor.
