+++ BEP.sln
@@ -10,7 +10,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Timetabling.Tests", "Timeta
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
 		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
@@ -20,22 +22,18 @@ Global
 		{2E9F822F-947E-43EE-A440-33C280785651}.Release|Any CPU.ActiveCfg = Release|x86
 		{2E9F822F-947E-43EE-A440-33C280785651}.Release|x86.ActiveCfg = Release|x86
 		{2E9F822F-947E-43EE-A440-33C280785651}.Release|x86.Build.0 = Release|x86
-		{84D69E31-7937-4D92-B5B5-377710C2EBF9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{84D69E31-7937-4D92-B5B5-377710C2EBF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{84D69E31-7937-4D92-B5B5-377710C2EBF9}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{84D69E31-7937-4D92-B5B5-377710C2EBF9}.Debug|x86.Build.0 = Debug|Any CPU
-		{84D69E31-7937-4D92-B5B5-377710C2EBF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{84D69E31-7937-4D92-B5B5-377710C2EBF9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{84D69E31-7937-4D92-B5B5-377710C2EBF9}.Release|x86.ActiveCfg = Release|Any CPU
-		{84D69E31-7937-4D92-B5B5-377710C2EBF9}.Release|x86.Build.0 = Release|Any CPU
-		{587B4DAD-9733-41EC-8ED3-5EA596FE6308}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{587B4DAD-9733-41EC-8ED3-5EA596FE6308}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{587B4DAD-9733-41EC-8ED3-5EA596FE6308}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{587B4DAD-9733-41EC-8ED3-5EA596FE6308}.Debug|x86.Build.0 = Debug|Any CPU
-		{587B4DAD-9733-41EC-8ED3-5EA596FE6308}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{587B4DAD-9733-41EC-8ED3-5EA596FE6308}.Release|Any CPU.Build.0 = Release|Any CPU
-		{587B4DAD-9733-41EC-8ED3-5EA596FE6308}.Release|x86.ActiveCfg = Release|Any CPU
-		{587B4DAD-9733-41EC-8ED3-5EA596FE6308}.Release|x86.Build.0 = Release|Any CPU
+		{84D69E31-7937-4D92-B5B5-377710C2EBF9}.Debug|Any CPU.ActiveCfg = Debug|x86
+		{84D69E31-7937-4D92-B5B5-377710C2EBF9}.Debug|x86.ActiveCfg = Debug|x86
+		{84D69E31-7937-4D92-B5B5-377710C2EBF9}.Debug|x86.Build.0 = Debug|x86
+		{84D69E31-7937-4D92-B5B5-377710C2EBF9}.Release|Any CPU.ActiveCfg = Release|x86
+		{84D69E31-7937-4D92-B5B5-377710C2EBF9}.Release|x86.ActiveCfg = Release|x86
+		{84D69E31-7937-4D92-B5B5-377710C2EBF9}.Release|x86.Build.0 = Release|x86
+		{587B4DAD-9733-41EC-8ED3-5EA596FE6308}.Debug|Any CPU.ActiveCfg = Debug|x86
+		{587B4DAD-9733-41EC-8ED3-5EA596FE6308}.Debug|x86.ActiveCfg = Debug|x86
+		{587B4DAD-9733-41EC-8ED3-5EA596FE6308}.Debug|x86.Build.0 = Debug|x86
+		{587B4DAD-9733-41EC-8ED3-5EA596FE6308}.Release|Any CPU.ActiveCfg = Release|x86
+		{587B4DAD-9733-41EC-8ED3-5EA596FE6308}.Release|x86.ActiveCfg = Release|x86
+		{587B4DAD-9733-41EC-8ED3-5EA596FE6308}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
