+++ Timetabling.Tests/packages.config
@@ -3,8 +3,14 @@
   <package id="Castle.Core" version="4.2.1" targetFramework="net461" />
   <package id="EntityFramework" version="6.2.0" targetFramework="net461" />
   <package id="Moq" version="4.8.2" targetFramework="net461" />
-  <package id="NUnit" version="3.10.1" targetFramework="net461" />
+  <package id="NUnit" version="3.5.0" targetFramework="net461" />
+  <package id="NUnit.Console" version="3.8.0" targetFramework="net461" />
   <package id="NUnit.ConsoleRunner" version="3.8.0" targetFramework="net461" />
+  <package id="NUnit.Extension.NUnitProjectLoader" version="3.6.0" targetFramework="net461" />
+  <package id="NUnit.Extension.NUnitV2Driver" version="3.7.0" targetFramework="net461" />
+  <package id="NUnit.Extension.NUnitV2ResultWriter" version="3.6.0" targetFramework="net461" />
+  <package id="NUnit.Extension.TeamCityEventListener" version="1.0.3" targetFramework="net461" />
+  <package id="NUnit.Extension.VSProjectLoader" version="3.7.0" targetFramework="net461" />
   <package id="System.IO.Abstractions" version="2.1.0.178" targetFramework="net461" />
   <package id="System.IO.Abstractions.TestingHelpers" version="2.1.0.178" targetFramework="net461" />
   <package id="System.Threading.Tasks.Extensions" version="4.4.0" targetFramework="net461" />
