+++ appveyor.yml
@@ -11,5 +11,5 @@ build:
   project: BEP.sln
   verbosity: normal
 test_script:
-  - packages\OpenCover.4.6.519\tools\OpenCover.Console.exe -register:user -target:"packages\NUnit.ConsoleRunner.3.8.0\tools\nunit3-console.exe" -targetargs:".\bin\Debug\BEP.exe --result=TestData.xml" -filter:"+[BEP]* -[BEP]*.Tests* -[BEP]*.DB*" -output:opencover.xml -returntargetcode
+  - packages\OpenCover.4.6.519\tools\OpenCover.Console.exe -register:user -target:"packages\NUnit.ConsoleRunner.3.8.0\tools\nunit3-console.exe" -targetargs:".\BEP.sln --result=TestData.xml --skipnontestassemblies" -filter:"+[Timetabling]*" -output:opencover.xml -returntargetcode
   - packages\coveralls.io.1.4.2\tools\coveralls.net.exe --opencover opencover.xml
